Lots of cleaning, Accelerometer and Music added
Cleaned up a lot of the code, adding comments, putting enums in
TetrisView, adding background(r,g,b) and fill(r.g.b) methods, merging
the draw methods, and more
Added a class to handle the accelerometer. Draws the current velocity
readings onto the screen, will be used as what seems to be a pretty
accurate gyroscope.
Put the music back in, still annoying and loud but not buggy with its
start/stops because it does that in onResume and onPause now